Major refactor
This refactor reworks both the console and cpu structs for 2 main reasons. The first is to separate out components such as ram and the rom file from the cpu struct. The second is to create a robust execution queue which can schedule events for the console to run on specified ticks.